+++ src/TodoScreen.js
@@ -1,15 +1,14 @@
 import React, { Component } from 'react';
 import { View } from 'react-native';
 
-import Footer from './components/Footer';
-import InputTodo from './components/InputTodo';
-import TodoList from './components/TodoList';
-import styles from './styles/base'; 
-import { todoNavOptions } from './constants/navigationOptions'
+import Footer from './components/Footer/Footer';
+import InputTodo from './components/InputTodo/InputTodo';
+import TodoList from './components/TodoList/TodoList';
+import styles from './styles/base';
+import { todoNavOptions } from './constants/navigationOptions';
 
-
-class TodoScreen extends Component { 
-  static navigationOptions = todoNavOptions
+class TodoScreen extends Component {
+  static navigationOptions = todoNavOptions;
 
   render() {
     return (
