+++ src/components/Footer/Footer.js
@@ -1,11 +1,11 @@
 import React, { Component } from 'react';
-import { View, Text } from 'react-native';
+import { View } from 'react-native';
 import { connect } from 'react-redux';
 
-import * as actionCreators from '../actions';
+import * as actionCreators from '../../actions';
+import { Button } from '../common/Button';
 
-import { Button } from './common/Button';
-import styles from './styles/footer';
+import styles from './styles';
 import { BUTTON_DELETE_TEXT } from './constants';
 
 const Footer = props => (
