+++ src/components/Footer/constants.js
@@ -1,0 +1,1 @@
+export const BUTTON_DELETE_TEXT = 'Remove Completed';
