+++ src/components/InputTodo/InputTodo.js
@@ -2,10 +2,10 @@ import React from 'react';
 import { View, TextInput } from 'react-native';
 import { connect } from 'react-redux';
 
-import * as actionCreators from '../actions';
+import * as actionCreators from '../../actions';
 
 import { PLACE_HOLDER } from './constants';
-import styles from './styles/inputTodo';
+import styles from './styles';
 
 const InputTodo = props => (
   <View style={styles.view}>
