+++ src/components/InputTodo/constants.js
@@ -1,0 +1,1 @@
+export const PLACE_HOLDER = 'Enter a item...';
