--- src/components/constants.js
@@ -1,3 +1,0 @@
-export const BUTTON_DELETE_TEXT = 'Remove Completed';
-export const HEADER_TEXT = 'Todo List';
-export const PLACE_HOLDER = 'Enter a item...';
